extend session logic
